Don't crash if sink has already been added in AddEncodedSink()
This CL allows VideoRtpTrackSource::AddEncodedSink(sink) to be called
with a sink that has been added before. It happens when the stream has
been cloned for instance. Following RemoveEncodedSink(sink) pattern,
it transforms a DCHECK in a conditional IF.

Note that this CL is needed for https://chromium-review.googlesource.com/c/chromium/src/+/4365020

No-Presubmit: true
Bug:chromium:1423413
Change-Id: I4a596883191bdb7f4fa947db21a44dca4f3bfa20