Re-land VP9 AllLayersInactive test by reverting is_svc tweak.
This test (previously landed here: [1]) got reverted in a standalone
CL because it was failing on the importer bot.

This CL attempts to re-land it by reverting a subset of the original CL
that may have contributed to the problem: [2].

The test does not misbehave locally or on the CQ, only on the importer,
so this is a speculative CL, but the changes to `is_svc` could be a
mistake.

[1] https://webrtc-review.googlesource.com/c/src/+/298881/5/pc/peer_connection_simulcast_unittest.cc
[2] https://webrtc-review.googlesource.com/c/src/+/298881/5/video/video_stream_encoder.cc

Bug: webrtc:15033
Change-Id: I2b7ffb5233c892f30aec933bbb7bcae4edfdbe5b