Add standard L1T3 test that all layers can be inactived.
A similar test to this was landed but got reverted due to failing on the
importer bot. That is being investigated here[1]. This CL relands parts
of the original test that we believe are NOT causing issues.

This CL is not expected to fail on the importer, let's land and verify.

[1] https://webrtc-review.googlesource.com/c/src/+/299006

Bug: webrtc:15033
Change-Id: If11e459b17d3606da3ddff0742e5429c642f7e83