Fix ICE fail issue when network status changes
Under GATHER_CONTINUALLY mode, when the network status changes (such as losing connection to WiFi, switching to cellular data after a period of time, losing connection again, switching back to WiFi or cellular data, and so on), ICE may enter a “fail” state and cannot recover to a connected state, which prevents the PC from continuing to work.

The process involves performing ICE regathering and ICE restart to resolve the issue.

Bug: None
Change-Id: I4c97c089c27a50d670222285eae842ef3bffa2bb