Introduce support for video packet batching.
This CL introduces a new class TransportSendBatchController, which
is injected into WebRTC via PeerConnectionInterface::RTCConfiguration.
The injected send batch controller is forwarded to the pacing
controller, which emits an OnSendBatchComplete call whenever a
burst of packets has been sent.

Furthermore, PacketOptions has been augmented with a "batchable"
attribute which gives injected AsyncPacketSockets a chance to
understand when a batch begins.

Bug: chromium:1439830
Change-Id: I1846b9d4a8a0efd227d617691213a2e048bdc8a2