Fixes distortion in WGC screen capture path
This is a partial revert of the previously landed CL in
https://webrtc-review.googlesource.com/c/src/+/301200.

I noticed when I used `getDisplayMedia` on my private Windows laptop
in combination with WGC that the captured screen was distorted and
did only contain tilted (~35 degrees) lines in all sorts of colors.

The issue only happened for one particular screen resolution
3000 x 2000 and 200% scale. Changing to 100% scaled resolved the
issue.

I tried many other resolutions but could only trigger for the one
above.

Next, I bisected and found [1] which led to [2] which contains my own
change in https://webrtc-review.googlesource.com/c/src/+/301200.

The only part that could affect the video frame was the part which
did `CopyPixelsFrom` so I reverted that part and it solved the issue
on my private Windows laptop.

I did not dig deeper into why this particular resolution triggered
the distortion but deiced to revert to avoid more reports.

[1] https://chromium.googlesource.com/chromium/src/+log/b4c2a7fcf591a52c7d07de92808047bf4e2da77d..ff848b7a4337fd22f5c52464bbec164e17b27fa3

[2] https://chromium.googlesource.com/chromium/src/+/a9a2957dbcfd9e2e670fa9a5241269888a0f55d2

Bug: chromium:1428592
Change-Id: I328e77840cd3ca6871254cdf06500bdc616b0c36