Change to use the RTCPReceivers own remote SSRC as key when saving RTT
When asking for RTT values from RTCPReceivers (e.g. useing LastRtt()), the remote_ssrc_ value is used to retrieve the information. So any RTT saved under a different key is lost to users of RTCPReceiver.

This change allow me to get Glass 2 Glass metric for Android to work, by giving a RTT value for video, so local to capture clock offset can be estimated.

Bug:None
Change-Id: If058cd8f0e34cd700fe3a7ac1fbe78f675abbf52