Improved event tracing in FrameCadenceAdapter
This CL is a minor cleanup of the work done earlier in [1].
We now remove the arguments which contained the measured delta times
between two TRACE_EVENT_ASYNC_BEGIN and TRACE_EVENT_ASYNC_END events.
Also, the ID for FrameToQueue1 is now unique which was not the case
previously.

The same information can be obtained from the `slice` table and the
`dur` key.

TODO: add measurement of FrameToQueue1 + FrameToQueue2.

[1] https://webrtc-review.googlesource.com/c/src/+/322121

Bug: webrtc:15456
Change-Id: Ibe2d7bb53380710671c2c36012dcd573942bae69