dcsctp: Pass max_timer_backoff_duration to timers
When this option was added[1], it was documented that it affected many
timers - not just t3-rtx. But for some reason, only t3-rtx was
configured with this option.

This CL adds it to t1-init, t1-cookie, t2-shutdown - as previously
documented - and also to the re-config timer.

This was found in a pass-by code review.

[1] https://webrtc-review.googlesource.com/c/src/+/230702

Bug: None
Change-Id: I4e6692e38472b1ca2997826ceb5a07253974bc0d