Remove SSRCs from libSRTP when removing them from the rtp_demuxer
This uses libSRTPs srtp_remove_stream()
  https://github.com/cisco/libsrtp/blob/main/include/srtp.h#L597
method to remove SSRCs from the libSRTP session when they are removed
from the RTP demuxer. This works even when the stream was added
automatically via the ssrc_any_inbound mechanism.

Manual test on
  https://webrtc.github.io/samples/src/content/peerconnection/bandwidth/
after making a call run
  pc1.getTransceivers()[0].stop();
  await pc1.setLocalDescription();
  await pc2.setRemoteDescription(pc1.localDescription);
which removes the ssrcs from the SDP. This will in turn remove
them from the demuxer and libSRTP.

BUG=TODO

Change-Id: I655bde5f8ddf26ac91395ef54bd1b3c598813380