[DO NOT LAND] Introduce vouchers.
This change introduces a new task queue concept, Voucher. They
are associated with a currently running task tree. Whenever
tasks are posted, the current voucher is inherited and set as
current to the new task.

The voucher exists for as long as there are direct and indirect
tasks running that descend from the task where the voucher was
created.

This particular change measures time from capture to all encode
operations complete. An accompanying Chrome change ensures
we track VEA operations in RTCVideoEncoder, and also includes the
first resulting packet sends.

Bug: None
Change-Id: I2a27800a4e5504f219d8b9d33c56a48904cf6dde